feat: 視点変換は共通なのでCameraUBOを使う

+++ webgl2/src/vertex.rs
@@ -82,7 +82,7 @@ impl VertexVbo {
             let view = js_sys::Float32Array::view(data);
             gl.buffer_sub_data_with_i32_and_array_buffer_view_and_src_offset_and_length(
                 Self::TARGET,
-                offset * P::size() as i32 * std::mem::size_of::<f32>() as i32,
+                offset * P::size() * std::mem::size_of::<f32>() as i32,
                 &view,
                 0,
                 P::size() as u32,
